ai-travel-agent/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
python-dotenv
requests
openai
pytest
